+++ pkg/config/config.go
@@ -42,7 +42,7 @@ type SMTP struct {
 	Domain          string        `required:"true" default:"inbucket" desc:"HELO domain"`
 	DomainNoStore   string        `desc:"Load testing domain"`
 	MaxRecipients   int           `required:"true" default:"200" desc:"Maximum RCPT TO per message"`
-	MaxMessageBytes int           `required:"true" default:"2048000" desc:"Maximum message size"`
+	MaxMessageBytes int           `required:"true" default:"10240000" desc:"Maximum message size"`
 	StoreMessages   bool          `required:"true" default:"true" desc:"Store incoming mail?"`
 	Timeout         time.Duration `required:"true" default:"300s" desc:"Idle network timeout"`
 }
@@ -57,10 +57,9 @@ type POP3 struct {
 // Web contains the HTTP server configuration.
 type Web struct {
 	Addr           string `required:"true" default:"0.0.0.0:9000" desc:"Web server IP4 host:port"`
-	TemplateDir    string `required:"true" default:"themes/bootstrap/templates" desc:"Theme template dir"`
+	UIDir          string `required:"true" default:"ui" desc:"User interface dir"`
+	GreetingFile   string `required:"true" default:"ui/greeting.html" desc:"Home page greeting HTML"`
 	TemplateCache  bool   `required:"true" default:"true" desc:"Cache templates after first use?"`
-	PublicDir      string `required:"true" default:"themes/bootstrap/public" desc:"Theme public dir"`
-	GreetingFile   string `required:"true" default:"themes/greeting.html" desc:"Home page greeting HTML"`
 	MailboxPrompt  string `required:"true" default:"@inbucket" desc:"Prompt next to mailbox input"`
 	CookieAuthKey  string `desc:"Session cipher key (text)"`
 	MonitorVisible bool   `required:"true" default:"true" desc:"Show monitor tab in UI?"`