+++ pkg/server/web/template.go
@@ -5,7 +5,6 @@ import (
 	"net/http"
 	"path"
 	"path/filepath"
-	"strings"
 	"sync"
 
 	"github.com/jhillyerd/inbucket/pkg/log"
@@ -49,8 +48,7 @@ func ParseTemplate(name string, partial bool) (*template.Template, error) {
 		return t, nil
 	}
 
-	tempPath := strings.Replace(name, "/", string(filepath.Separator), -1)
-	tempFile := filepath.Join(rootConfig.Web.TemplateDir, tempPath)
+	tempFile := filepath.Join(rootConfig.Web.UIDir, templateDir, filepath.FromSlash(name))
 	log.Tracef("Parsing template %v", tempFile)
 
 	var err error
@@ -62,7 +60,8 @@ func ParseTemplate(name string, partial bool) (*template.Template, error) {
 		t, err = t.ParseFiles(tempFile)
 	} else {
 		t = template.New("_base.html").Funcs(TemplateFuncs)
-		t, err = t.ParseFiles(filepath.Join(rootConfig.Web.TemplateDir, "_base.html"), tempFile)
+		t, err = t.ParseFiles(
+			filepath.Join(rootConfig.Web.UIDir, templateDir, "_base.html"), tempFile)
 	}
 	if err != nil {
 		return nil, err